+++ lte/gateway/c/session_manager/test/test_session_state_5g.cpp
@@ -22,118 +22,20 @@
 #include "ProtobufCreators.h"
 #include "SessiondMocks.h"
 #include "SessionState.h"
-#include "SessionStateTester.h"
+#include "SessionStateTester5g.h"
 
 using ::testing::Test;
 
 namespace magma {
 
-class SessionStateTest5G : public ::testing::Test {
- protected:
-  virtual void SetUp() {
-    Teids teids;
-    rule_store = std::make_shared<StaticRuleStore>();
-    cfg        = build_sm_context(IMSI1, "10.20.30.40", 5);
-    session_state =
-        std::make_shared<SessionState>(IMSI1, SESSION_ID_1, cfg, *rule_store);
-    session_state->set_fsm_state(SESSION_ACTIVE, nullptr);
-    session_state->set_create_session_response(
-        CreateSessionResponse(), nullptr);
-    update_criteria = get_default_update_criteria();
-  }
-
-  void insert_static_rule_into_store(
-      uint32_t rating_group, const std::string& m_key,
-      const std::string& rule_id) {
-    rule_store->insert_rule(create_policy_rule(rule_id, m_key, rating_group));
-  }
-
-  void insert_static_rule_with_qos_into_store(
-      uint32_t rating_group, const std::string& m_key, const int qci,
-      const std::string& rule_id) {
-    PolicyRule rule =
-        create_policy_rule_with_qos(rule_id, m_key, rating_group, qci);
-    rule_store->insert_rule(rule);
-  }
-
-  uint32_t get_monitored_rule_count(const std::string& mkey) {
-    std::vector<PolicyRule> rules;
-    EXPECT_TRUE(session_state->get_dynamic_rules().get_rules(rules));
-    uint32_t count = 0;
-    for (PolicyRule& rule : rules) {
-      if (rule.monitoring_key() == mkey) {
-        count++;
-      }
-    }
-    return count;
-  }
-
-  SessionConfig build_sm_context(
-      const std::string& imsi,  // assumes IMSI prefix
-      const std::string& ue_ipv4, uint32_t pdu_id) {
-    SetSMSessionContext request;
-    auto* req =
-        request.mutable_rat_specific_context()->mutable_m5gsm_session_context();
-    auto* reqcmn = request.mutable_common_context();
-    req->set_pdu_session_id(pdu_id);
-    req->set_request_type(magma::RequestType::INITIAL_REQUEST);
-    req->mutable_pdu_address()->set_redirect_address_type(
-        magma::RedirectServer::IPV4);
-    req->mutable_pdu_address()->set_redirect_server_address(ue_ipv4);
-    req->set_priority_access(magma::priorityaccess::High);
-    req->set_imei("123456789012345");
-    req->set_gpsi("9876543210");
-    req->set_pcf_id("1357924680123456");
-
-    reqcmn->mutable_sid()->set_id(imsi);
-    reqcmn->set_sm_session_state(magma::SMSessionFSMState::CREATING_0);
-
-    SessionConfig cfg;
-    cfg.common_context       = request.common_context();
-    cfg.rat_specific_context = request.rat_specific_context();
-    cfg.rat_specific_context.mutable_m5gsm_session_context()->set_ssc_mode(
-        SSC_MODE_3);
-    return cfg;
-  }
-
-  uint32_t activate_5g_rule(
-      uint32_t rating_group, const std::string& m_key,
-      const std::string& rule_id, PolicyType rule_type,
-      std::time_t activation_time, std::time_t deactivation_time) {
-    PolicyRule rule = create_policy_rule(rule_id, m_key, rating_group);
-    RuleLifetime lifetime(activation_time, deactivation_time);
-    switch (rule_type) {
-      case STATIC:
-        rule_store->insert_rule(rule);
-        return session_state
-            ->activate_static_5g_rule(rule_id, lifetime, &update_criteria)
-            .version;
-        break;
-      case DYNAMIC:
-        return session_state
-            ->insert_dynamic_5g_rule(rule, lifetime, &update_criteria)
-            .version;
-        break;
-      default:
-        break;
-    }
-    return 0;
-  }
-
- protected:
-  std::shared_ptr<StaticRuleStore> rule_store;
-  std::shared_ptr<SessionState> session_state;
-  SessionStateUpdateCriteria update_criteria;
-  SessionConfig cfg;
-};
-
 TEST_F(SessionStateTest5G, test_session_rules) {
-  activate_5g_rule(1, "m1", "rule1", DYNAMIC, 0, 0);
+  session_state->set_config(cfg, nullptr);
+  activate_rule(1, "m1", "rule1", DYNAMIC, 0, 0);
   EXPECT_EQ(1, session_state->total_monitored_rules_count());
-  activate_5g_rule(2, "m2", "rule2", STATIC, 0, 0);
+  activate_rule(2, "m2", "rule2", STATIC, 0, 0);
   EXPECT_EQ(2, session_state->total_monitored_rules_count());
   // add a OCS-ONLY static rule
-  activate_5g_rule(3, "", "rule3", STATIC, 0, 0);
+  activate_rule(3, "", "rule3", STATIC, 0, 0);
   EXPECT_EQ(2, session_state->total_monitored_rules_count());
 
   std::vector<std::string> rules_out{};
@@ -152,10 +54,10 @@ TEST_F(SessionStateTest5G, test_session_rules) {
 
   // Test rule removals
   PolicyRule rule_out;
-  session_state->deactivate_static_5g_rule("rule2", &update_criteria);
+  session_state->deactivate_static_rule("rule2", &update_criteria);
   EXPECT_EQ(1, session_state->total_monitored_rules_count());
-  EXPECT_TRUE(session_state->remove_dynamic_5g_rule(
-      "rule1", &rule_out, &update_criteria));
+  EXPECT_TRUE(
+      session_state->remove_dynamic_rule("rule1", &rule_out, &update_criteria));
   EXPECT_EQ("m1", rule_out.monitoring_key());
   EXPECT_EQ(0, session_state->total_monitored_rules_count());
 
@@ -178,11 +80,12 @@ TEST_F(SessionStateTest5G, test_session_rules) {
 }
 
 TEST_F(SessionStateTest5G, test_get_session_rules) {
+  session_state->set_config(cfg, nullptr);
   // populate rule store with 2 static and 2 dynamic rules
-  activate_5g_rule(1, "", "rule-static-1", STATIC, 0, 0);
-  activate_5g_rule(2, "m1", "rule-static-2", STATIC, 0, 0);
-  activate_5g_rule(1, "", "rule-dynamic-1", DYNAMIC, 0, 0);
-  activate_5g_rule(2, "m1", "rule-dynamic-2", DYNAMIC, 0, 0);
+  activate_rule(1, "", "rule-static-1", STATIC, 0, 0);
+  activate_rule(2, "m1", "rule-static-2", STATIC, 0, 0);
+  activate_rule(1, "", "rule-dynamic-1", DYNAMIC, 0, 0);
+  activate_rule(2, "m1", "rule-dynamic-2", DYNAMIC, 0, 0);
 
   EXPECT_TRUE(session_state->is_static_rule_installed("rule-static-1"));
   EXPECT_TRUE(session_state->is_static_rule_installed("rule-static-2"));
@@ -238,6 +141,7 @@ TEST_F(SessionStateTest5G, test_get_session_rules) {
 }
 
 TEST_F(SessionStateTest5G, test_process_static_rule_installs) {
+  session_state->set_config(cfg, nullptr);
   // Insert 2 static rules without qos into static rule store
   insert_static_rule_into_store(0, "mkey1", "static-1");
   insert_static_rule_into_store(0, "mkey1", "static-2");
@@ -247,8 +151,8 @@ TEST_F(SessionStateTest5G, test_process_static_rule_installs) {
 
   // activate static-1 and static-qos-3 in advance
   RuleLifetime lifetime;
-  session_state->activate_static_5g_rule("static-1", lifetime, nullptr);
-  session_state->activate_static_5g_rule("static-qos-3", lifetime, nullptr);
+  session_state->activate_static_rule("static-1", lifetime, nullptr);
+  session_state->activate_static_rule("static-qos-3", lifetime, nullptr);
 
   // Create a StaticRuleInstall with all four rules above
   std::vector<StaticRuleInstall> rule_installs{
@@ -263,9 +167,10 @@ TEST_F(SessionStateTest5G, test_process_static_rule_installs) {
   };
   EXPECT_EQ(4, rule_installs.size());
   RulesToProcess pending_activation, pending_deactivation;
-  session_state->process_static_5g_rule_installs(
-      rule_installs, &pending_activation, &pending_deactivation,
-      &update_criteria);
+  RulesToSchedule pending_scheduling;
+  session_state->process_static_rule_installs(
+      rule_installs, &pending_activation, &pending_deactivation, nullptr,
+      &pending_scheduling, &update_criteria);
   EXPECT_EQ(2, pending_activation.size());
   EXPECT_EQ("static-2", pending_activation[0].rule.id());
   EXPECT_EQ(1, pending_activation[0].version);
@@ -278,6 +183,7 @@ TEST_F(SessionStateTest5G, test_process_static_rule_installs) {
 }
 
 TEST_F(SessionStateTest5G, test_process_dynamic_rule_installs) {
+  session_state->set_config(cfg, nullptr);
   PolicyRule dynamic_1 = create_policy_rule("dynamic-1", "", 0);
   PolicyRule dynamic_2 = create_policy_rule("dynamic-2", "", 0);
   PolicyRule dynamic_qos_3 =
@@ -288,8 +194,8 @@ TEST_F(SessionStateTest5G, test_process_dynamic_rule_installs) {
   // Install dynamic rules for dynamic-1 and dynamic-qos-3
   // Then version should be increased.
   RuleLifetime lifetime;
-  session_state->insert_dynamic_5g_rule(dynamic_1, lifetime, nullptr);
-  session_state->insert_dynamic_5g_rule(dynamic_qos_3, lifetime, nullptr);
+  session_state->insert_dynamic_rule(dynamic_1, lifetime, nullptr);
+  session_state->insert_dynamic_rule(dynamic_qos_3, lifetime, nullptr);
 
   // Create a StaticRuleInstall with all four rules above
   std::vector<DynamicRuleInstall> rule_installs{
@@ -304,9 +210,10 @@ TEST_F(SessionStateTest5G, test_process_dynamic_rule_installs) {
   };
   EXPECT_EQ(4, rule_installs.size());
   RulesToProcess pending_activation, pending_deactivation;
-  session_state->process_dynamic_5g_rule_installs(
-      rule_installs, &pending_activation, &pending_deactivation,
-      &update_criteria);
+  RulesToSchedule pending_scheduling;
+  session_state->process_dynamic_rule_installs(
+      rule_installs, &pending_activation, &pending_deactivation, nullptr,
+      &pending_scheduling, &update_criteria);
   EXPECT_EQ(4, pending_activation.size());
   EXPECT_EQ("dynamic-1", pending_activation[0].rule.id());
   EXPECT_EQ(2, pending_activation[0].version);
@@ -325,6 +232,7 @@ TEST_F(SessionStateTest5G, test_process_dynamic_rule_installs) {
 }
 
 TEST_F(SessionStateTest5G, test_remove_all_session_rules) {
+  session_state->set_config(cfg, nullptr);
   // Insert 1 static rules without qos into static rule store
   insert_static_rule_into_store(0, "mkey1", "static-1");
   // Insert 1 static rules with qos into static rule store
@@ -348,9 +256,10 @@ TEST_F(SessionStateTest5G, test_remove_all_session_rules) {
   EXPECT_EQ(2, dynamic_rule_installs.size());
 
   RulesToProcess pending_activation, pending_deactivation;
-  session_state->process_5g_rules_to_install(
+  RulesToSchedule pending_scheduling;
+  session_state->process_rules_to_install(
       static_rule_installs, dynamic_rule_installs, &pending_activation,
-      &pending_deactivation, &update_criteria);
+      &pending_deactivation, nullptr, &pending_scheduling, &update_criteria);
 
   EXPECT_EQ(4, pending_activation.size());
 
@@ -371,7 +280,7 @@ TEST_F(SessionStateTest5G, test_remove_all_session_rules) {
   EXPECT_EQ("dynamic-1", update_criteria.dynamic_rules_to_install[0].id());
   EXPECT_EQ("dynamic-2", update_criteria.dynamic_rules_to_install[1].id());
 
-  session_state->remove_all_5g_rules_for_termination(&update_criteria);
+  session_state->remove_all_rules_for_termination(&update_criteria);
 
   EXPECT_EQ(update_criteria.static_rules_to_uninstall.size(), 1);
   EXPECT_EQ(update_criteria.dynamic_rules_to_uninstall.size(), 2);
